auth first step

+++ internal/pkg/filter/usecase/usecase.go
@@ -26,6 +26,7 @@ func (f *FilterUseCase) AddFilters(userId uint, filterInp filter.FilterInput) er
 	if err != nil {
 		return err
 	}
+
 	var addReason []dataStruct.UserReason
 	for _, reason := range reasonId {
 		var userReason dataStruct.UserReason
@@ -93,10 +94,11 @@ func (f *FilterUseCase) ChangeFilters(userId uint, filterInp filter.FilterInput)
 			addReasons = append(addReasons, reasonAdd)
 		}
 	}
-
-	err = f.Repo.AddUserReason(addReasons)
-	if err != nil {
-		return err
+	if len(addReasons) > 0 {
+		err = f.Repo.AddUserReason(addReasons)
+		if err != nil {
+			return err
+		}
 	}
 
 	var deleteReason []uint
@@ -105,9 +107,12 @@ func (f *FilterUseCase) ChangeFilters(userId uint, filterInp filter.FilterInput)
 			deleteReason = append(deleteReason, reason)
 		}
 	}
-	err = f.Repo.DeleteUserReason(userId, deleteReason)
-	if err != nil {
-		return err
+
+	if len(deleteReason) > 0 {
+		err = f.Repo.DeleteUserReason(userId, deleteReason)
+		if err != nil {
+			return err
+		}
 	}
 
 	var newFilter dataStruct.UserFilter
